Keep refreshing properties when paused or blank
Continue polling even when the player isn't running.
This doesn't make sense if Kodi events are available, but if events
don't work, it makes Kaidi recognize when another client starts playback
or switches tracks while playback isn't running.
In conjunction with stopping polling when the document is invisible, it
should have little cost.